refactor: LLM 文档生成插件抽离为公共包 packages/common-docs
- 新增 packages/common-docs/plugins/generate-llms 共享插件
  （types/markdown 清洗/插件主体/README），支持 componentMap、
  componentsDir、readDemoCode 等配置项
- miniprogram / chat / uniapp 三站点改为薄接入层，仅传配置
- 后续 vue-next / react / mobile-vue 站点可直接复用
